# More aggressive parallelization for BCn encoding