This PR lowers the group sizes required for images to be split when encoding with BCn.

Other changes:
- Fixed a bug in `PreferredFragmentSize::new` where log2 wasn't calculated correctly.
- Changed terminology of "group size" to "fragment size" when talking about split views. This caused a lot of code changes, but makes the code easier to understand.

## Background

Parallelization comes with overhead, so I decided to only parallelize images that are large enough to make it worth it. This is done via `SplitView`. `SplitView` splits an image into multiple independent fragments that together make up the whole image. Currently, BCn formats define a minimum size for these fragments by requiring that they contain a certain number of pixels. E.g. BC1 with High quality requires >=256 pixels per fragment. This is called the preferred fragment size.

This system ensures that small images aren't needlessly parallelized. The minimum size for parallelization can be controlled using the preferred fragment size.

## Benchmark

I added a new benchmark to measure the time it takes to encode a single surface of a given size with a given format and quality, with and without parallelization. I compared (1) single-threaded performance, (2) parallelization with old fragment sizes, and (3) parallelization with new fragment sizes.

(Times are with PR #72 already applied.)

| Format    | Quality | Image size | One thread |   Old par |   New par | Speed up |
| --------- | ------- | ---------- | ---------: | --------: | --------: | -------: |
| BC1_UNORM | Fast    | 64x64      |    0.23 ms |   0.23 ms |   0.22 ms |          |
|           |         | 128x128    |    0.94 ms |   0.92 ms |   0.38 ms |     2.4x |
|           |         | 256x256    |    3.75 ms |   3.65 ms |   0.74 ms |     4.9x |
|           |         | 512x512    |   14.99 ms |   7.60 ms |   2.24 ms |     3.4x |
|           |         | 1024x1024  |   60.03 ms |  11.78 ms |   8.43 ms |     1.4x |
|           | Normal  | 64x64      |    1.19 ms |   1.17 ms |   0.48 ms |     2.4x |
|           |         | 128x128    |    4.73 ms |   4.61 ms |   0.99 ms |     4.7x |
|           |         | 256x256    |   19.01 ms |   9.66 ms |   2.96 ms |     3.3x |
|           |         | 512x512    |   75.72 ms |  15.44 ms |  10.96 ms |     1.4x |
|           |         | 1024x1024  |  304.36 ms |  46.38 ms |  42.94 ms |          |
|           | High    | 64x64      |    3.33 ms |   3.28 ms |   0.70 ms |     4.7x |
|           |         | 128x128    |   13.38 ms |   6.80 ms |   2.11 ms |     3.2x |
|           |         | 256x256    |   52.89 ms |  10.94 ms |   8.14 ms |     1.3x |
|           |         | 512x512    |  213.35 ms |  32.68 ms |  30.75 ms |          |
|           |         | 1024x1024  |  853.56 ms | 122.18 ms | 123.64 ms |          |
|           |         |            |            |           |           |          |
| BC4_UNORM | Fast    | 64x64      |    0.16 ms |   0.15 ms |   0.15 ms |          |
|           |         | 128x128    |    0.65 ms |   0.64 ms |   0.27 ms |     2.4x |
|           |         | 256x256    |    2.59 ms |   2.52 ms |   0.54 ms |     4.7x |
|           |         | 512x512    |   10.41 ms |  10.41 ms |   1.58 ms |     6.6x |
|           |         | 1024x1024  |   42.05 ms |  12.59 ms |   5.95 ms |     2.1x |
|           | Normal  | 64x64      |    0.34 ms |   0.34 ms |   0.22 ms |     1.5x |
|           |         | 128x128    |    1.42 ms |   1.41 ms |   0.32 ms |     4.4x |
|           |         | 256x256    |    5.58 ms |   5.60 ms |   0.90 ms |     6.2x |
|           |         | 512x512    |   22.77 ms |   7.56 ms |   3.28 ms |     2.3x |
|           |         | 1024x1024  |   91.85 ms |  16.97 ms |  12.87 ms |     1.3x |
|           | High    | 64x64      |    0.75 ms |   0.75 ms |   0.31 ms |     2.4x |
|           |         | 128x128    |    3.02 ms |   2.99 ms |   0.63 ms |     4.7x |
|           |         | 256x256    |   12.39 ms |   6.17 ms |   1.85 ms |     3.3x |
|           |         | 512x512    |   51.21 ms |   9.79 ms |   7.33 ms |     1.3x |
|           |         | 1024x1024  |  197.81 ms |  29.23 ms |  26.89 ms |          |

Note that:

1. There are no regressions (within the margin of error). Parallel encoding is now as fast as before, or faster.
2. There are no major changes to large images (>=1k). This is to be expected since they were already decently parallelized.
3. Medium-sized images (128-512) improved the most.